Revised the Cluster List
Revised it so that the inner cluster also checks if the CURRENT tokenized method identifier is not already in the inner cluster. If the identifier is not already in the inner cluster, then add it. Furthermore the final line of code is a print statement to test the cluster list, to use it just uncomment it.